refactor(bayn): modularize paper store
